+++ src/Config/index.ts
@@ -1,0 +1,7 @@
+const Config = {
+    version: "1.0.2",
+    releasedate: "2024-12-08",
+    visiblebranch: "Development",
+}
+
+export { Config };