+++ src/app/Home/page.tsx
@@ -15,8 +15,12 @@ import {
     Language,
     LightMode,
     DarkMode,
-    Contrast
+    Contrast,
+    Build,
+    Bookmark,
+    QrCode2,
 } from '@mui/icons-material';
+import Link from "next/link";
 
 let ProjectInfo = {
     version: "1.0.2",
@@ -63,6 +67,26 @@ const Links: Array<{
                 </svg>
             ),
             isEnable: true,
+        },
+        {
+            Label: "Discord",
+            URL: "https://discord.gg/3j2Zt5fU",
+            Component: (
+                <svg xmlns="http://www.w3.org/2000/svg" width="23" height="23" fill="currentColor" viewBox="0 0 16 16">
+                    <path d="M13.545 2.907a13.227 13.227 0 0 0-3.257-1.011.05.05 0 0 0-.052.025c-.141.25-.297.577-.406.833a12.19 12.19 0 0 0-3.658 0 8.258 8.258 0 0 0-.412-.833.051.051 0 0 0-.052-.025c-1.125.194-2.22.534-3.257 1.011a.041.041 0 0 0-.021.018C.356 6.024-.213 9.047.066 12.032c.001.014.01.028.021.037a13.276 13.276 0 0 0 3.995 2.02.05.05 0 0 0 .056-.019c.308-.42.582-.863.818-1.329a.05.05 0 0 0-.01-.059.051.051 0 0 0-.018-.011 8.875 8.875 0 0 1-1.248-.595.05.05 0 0 1-.02-.066.051.051 0 0 1 .015-.019c.084-.063.168-.129.248-.195a.05.05 0 0 1 .051-.007c2.619 1.196 5.454 1.196 8.041 0a.052.052 0 0 1 .053.007c.08.066.164.132.248.195a.051.051 0 0 1-.004.085 8.254 8.254 0 0 1-1.249.594.05.05 0 0 0-.03.03.052.052 0 0 0 .003.041c.24.465.515.909.817 1.329a.05.05 0 0 0 .056.019 13.235 13.235 0 0 0 4.001-2.02.049.049 0 0 0 .021-.037c.334-3.451-.559-6.449-2.366-9.106a.034.034 0 0 0-.02-.019Zm-8.198 7.307c-.789 0-1.438-.724-1.438-1.612 0-.889.637-1.613 1.438-1.613.807 0 1.45.73 1.438 1.613 0 .888-.637 1.612-1.438 1.612Zm5.316 0c-.788 0-1.438-.724-1.438-1.612 0-.889.637-1.613 1.438-1.613.807 0 1.451.73 1.438 1.613 0 .888-.631 1.612-1.438 1.612Z" />
+                </svg>
+            ),
+            isEnable: true,
+        },
+        {
+            Label: "Mail",
+            URL: "mailto:abc@gmail.com",
+            Component: (
+                <svg xmlns="http://www.w3.org/2000/svg" className="stackoverflowsvg" width="32" height="32" role="img" viewBox="0 0 24 24">
+                    <path d="M24 5.457v13.909c0 .904-.732 1.636-1.636 1.636h-3.819V11.73L12 16.64l-6.545-4.91v9.273H1.636A1.636 1.636 0 0 1 0 19.366V5.457c0-2.023 2.309-3.178 3.927-1.964L5.455 4.64 12 9.548l6.545-4.91 1.528-1.145C21.69 2.28 24 3.434 24 5.457z" />
+                </svg>
+            ),
+            isEnable: true,
         }
     ];
 
@@ -101,13 +125,48 @@ export default function Home() {
                         if (item.isEnable) {
                             return (
                                 <Tooltip key={index} content={item.Label} placement="top">
-                                    <a href={item.URL}>
+                                    <a href={item.URL} target="_blank">
                                         {item.Component}
                                     </a>
                                 </Tooltip>
                             );
                         }
                     })}
+                </motion.div>
+            </motion.div>
+            <motion.div
+                className="ItemRow"
+                initial={{ opacity: 0, scale: 0.9 }}
+                animate={{ opacity: 1, scale: 1 }}
+                transition={{ delay: 0.2, duration: 0.7, ease: "easeInOut" }}
+            >
+                <motion.div
+                    className="Item"
+                    initial={{ opacity: 0, scale: 0.9 }}
+                    animate={{ opacity: 1, scale: 1 }}
+                    transition={{ delay: 0.2, duration: 0.7, ease: "easeInOut" }}
+                >
+                    <Link href={"/Tools"} className="flex row gap-4 items-center">
+                        <Bookmark sx={{
+                            width: 32,
+                            height: 32,
+                        }} />
+                        Bookmarks
+                    </Link>
+                </motion.div>
+                <motion.div
+                    className="Item"
+                    initial={{ opacity: 0, scale: 0.9 }}
+                    animate={{ opacity: 1, scale: 1 }}
+                    transition={{ delay: 0.2, duration: 0.7, ease: "easeInOut" }}
+                >
+                    <Link href={"/Tools/QR"} className="flex row gap-4 items-center max-h max-w">
+                        <QrCode2 sx={{
+                            width: 32,
+                            height: 32,
+                        }} />
+                        QR
+                    </Link>
                 </motion.div>
 
             </motion.div>